LocationService works but at the moment you need to reopen the NearMeFragment in order to see the positions (to be fixed)

+++ app/src/main/java/it/unipd/dei/esp/whatsapd/ui/nearme/PoiWrapper.kt
@@ -9,7 +9,7 @@ import it.unipd.dei.esp.whatsapd.repository.database.Poi
  */
 
 class PoiWrapper(
-	poi: Poi, val distance: Double
+	poi: Poi, val distance: Int
 ) : Poi(
 	poi.name,
 	poi.latitude,
